requirements.txt
fastapi>=0.109.0
pydantic>=2.6.0
httpx>=0.26.0
python-dotenv>=1.0.0
uvicorn>=0.27.0
requests>=2.31.0
notion-client>=2.2.1
openai>=1.3.0